move IterableLikeToIterableTransformer to package typeutils
