+++ logic/atrium-logic-common/src/main/kotlin/ch/tutteli/atrium/logic/utils/varArgHelpers.kt
@@ -5,7 +5,8 @@ import ch.tutteli.atrium.creating.AssertionContainer
 import ch.tutteli.atrium.logic.creating.charsequence.contains.CharSequenceContains
 import ch.tutteli.atrium.logic.creating.iterable.contains.IterableLikeContains
 import ch.tutteli.atrium.logic.creating.typeutils.IterableLike
-import ch.tutteli.atrium.logic.utils.impl.DefaultIterableLikeToIterableTransformer
+import ch.tutteli.atrium.logic.creating.typeutils.IterableLikeToIterableTransformer
+import ch.tutteli.atrium.logic.creating.typeutils.impl.DefaultIterableLikeToIterableTransformer
 
 /**
  * Transforms the given [IterableLike] to `Pair<T, Array<out T>>` with the intend that it can be easily used for a function
